Update configuration structure for no_transport
