+++ seapopym/configuration/no_transport/environment_parameter.py
@@ -2,17 +2,25 @@
 
 from __future__ import annotations
 
-from typing import Literal
+from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 from attrs import define, field, frozen, validators
 from dask.distributed import Client
 
+from seapopym.configuration.abstract_configuration import (
+    AbstractChunkParameter,
+    AbstractClientParameter,
+    AbstractEnvironmentParameter,
+)
 from seapopym.logging.custom_logger import logger
+
+if TYPE_CHECKING:
+    from numbers import Number
 
 
 @frozen
-class ChunkParameter:
+class ChunkParameter(AbstractChunkParameter):
     """The chunk size of the different dimensions."""
 
     functional_group: Literal["auto"] | int | None = field(
@@ -54,7 +62,7 @@ class ChunkParameter:
 
 
 @define
-class ClientParameter:
+class ClientParameter(AbstractClientParameter):
     """
     The client parameter for the Dask client.
 
@@ -73,9 +81,7 @@ class ClientParameter:
         validator=validators.optional(validators.instance_of(int)),
         metadata={"description": "The number of threads per worker."},
     )
-    memory_limit: str | int | float = field(
-        default="auto", metadata={"description": "The memory limit of each worker."}
-    )
+    memory_limit: str | Number = field(default="auto", metadata={"description": "The memory limit of each worker."})
     client: Client | None = field(init=False, default=None, metadata={"description": "The Dask client."})
 
     @classmethod
@@ -133,83 +139,8 @@ class ClientParameter:
         self.client = None
 
 
-# @frozen(kw_only=True)
-# class BaseOuputForcingParameter:
-#     """A base class for the output forcing parameter."""
-
-#     path: str | Path = field(
-#         default=Path("./output.nc"),
-#         converter=Path,
-#         metadata={"description": "The path where the forcing will be stored."},
-#     )
-#     with_parameter: bool = field(
-#         default=True,
-#         validator=validators.instance_of(bool),
-#         metadata={"description": "If True, the forcing will be computed with the parameter forcing."},
-#     )
-#     with_forcing: bool = field(
-#         default=False,
-#         validator=validators.instance_of(bool),
-#         metadata={"description": "If True, forcing will be added to the output dataset."},
-#     )
-
-
-# @frozen(kw_only=True)
-# class BiomassParameter(BaseOuputForcingParameter):
-#     """The output parameter for the biomass forcing."""
-
-
-# @frozen(kw_only=True)
-# class ProductionParameter(BaseOuputForcingParameter):
-#     """The output parameter for the production forcing."""
-
-
-# @frozen(kw_only=True)
-# class PreProductionParameter(BaseOuputForcingParameter):
-#     """The output parameter for the pre-production forcing (i.e. with cohorts)."""
-
-#     timestamps: Iterable[str] | Iterable[int] | Literal["all"] = field(
-#         default=[-1],
-#         converter=lambda x: [x] if x != "all" and isinstance(x, (int, str)) else x,
-#         metadata={"description": "The timestamps for the pre-production forcing."},
-#     )
-
-#     @timestamps.validator
-#     def _validate_timestamps(
-#         self: PreProductionParameter, attribute: str, value: Iterable[str] | Iterable[int]
-#     ) -> None:
-#         if value == "all":
-#             return
-
-#         msg = "The timestamps must be either 'all' or a list of integers (time index) or a list of strings (datetime)."
-#         if not isinstance(value, Iterable):
-#             raise TypeError(msg)
-#         if all(isinstance(x, int) for x in value):
-#             return
-#         if all(isinstance(x, str) for x in value):
-#             return
-#         raise TypeError(msg)
-
-
-# @frozen(kw_only=True)
-# class OutputParameter:
-#     """The output parameter that manage the backup of the output forcings."""
-
-#     biomass: BiomassParameter = field(
-#         factory=BiomassParameter,
-#         validator=validators.instance_of(BiomassParameter),
-#         metadata={"description": "The output parameter for the biomass forcing."},
-#     )
-#     pre_production: PreProductionParameter = field(
-#         default=None,
-#         validator=validators.optional(validators.instance_of(PreProductionParameter)),
-#         metadata={"description": "The output parameter for the pre-production forcing."},
-#     )
-#     initial_conditions:
-
-
 @frozen(kw_only=True)
-class EnvironmentParameter:
+class EnvironmentParameter(AbstractEnvironmentParameter):
     """Manage the different environment parameters of the simulation."""
 
     chunk: ChunkParameter = field(
@@ -222,8 +153,3 @@ class EnvironmentParameter:
         validator=validators.instance_of(ClientParameter),
         metadata={"description": "The client parameter."},
     )
-    # output: OutputParameter = field(
-    #     factory=OutputParameter,
-    #     validator=validators.instance_of(OutputParameter),
-    #     metadata={"description": "The output parameter."},
-    # )
