+++ seapopym/configuration/no_transport/forcing_parameter.py
@@ -2,18 +2,21 @@
 
 from __future__ import annotations
 
+from collections.abc import Iterable
+from numbers import Number
 from pathlib import Path
-from typing import Iterable, Literal, ParamSpecArgs, ParamSpecKwargs
+from typing import Literal, ParamSpecArgs, ParamSpecKwargs
 
 import cf_xarray  # noqa: F401
 import fsspec
 import numpy as np
 import pandas as pd
 import xarray as xr
-from attrs import converters, field, frozen
+from attrs import converters, field, frozen, validators
 
+from seapopym.exception.parameter_exception import DifferentForcingTimestepError
 from seapopym.logging.custom_logger import logger
-from seapopym.standard.units import check_units
+from seapopym.standard.units import StandardUnitsLabels, check_units
 
 DECIMALS = 5  # ie. 1e-5 degrees which is equivalent to ~1m at the equator
 
@@ -104,7 +107,7 @@ class ForcingUnit:
         metadata={"description": "Forcing field."},
     )
 
-    resolution: Iterable[float, float] | float | None = field(
+    resolution: Iterable[Number, Number] | Number | None = field(
         default=None,
         converter=converters.optional(
             lambda x: tuple(float(item) for item in x) if isinstance(x, Iterable) else float(x)
@@ -126,7 +129,7 @@ class ForcingUnit:
         cls: ForcingUnit,
         forcing: xr.Dataset,
         name: str,
-        resolution: tuple[float, float] | float | None,
+        resolution: tuple[Number, Number] | Number | None,
         timestep: int | None,
     ) -> ForcingUnit:
         """Create a ForcingUnit from a path and a name."""
@@ -148,7 +151,7 @@ class ForcingUnit:
         cls: ForcingUnit,
         forcing: Path | str,
         name: str,
-        resolution: tuple[float, float] | float | None = None,
+        resolution: tuple[Number, Number] | Number | None = None,
         timestep: int | None = None,
         engine: Literal["zarr", "netcdf"] = "zarr",
         *args: ParamSpecArgs,
@@ -157,7 +160,6 @@ class ForcingUnit:
         """Create a ForcingUnit from a path and a name."""
         path_validation(forcing)
         data = xr.open_dataset(forcing, *args, engine=engine, **kwargs)
-        # data = xr.open_zarr(forcing) if engine == "zarr" else xr.open_dataset(forcing)
         return cls.from_dataset(data, name, resolution, timestep)
 
     def __attrs_post_init__(self: ForcingUnit) -> None:
@@ -180,3 +182,117 @@ class ForcingUnit:
             object.__setattr__(self, "forcing", check_units(self.forcing, units))
             return self
         return ForcingUnit(forcing=check_units(self.forcing, units), resolution=self.resolution, timestep=self.timestep)
+
+
+@frozen(kw_only=True)
+class ForcingParameters:
+    """
+    This data class is used to store access paths to forcing fields. You can inherit it to add further forcings, but in
+    this case you'll need to add new behaviors to the functions and classes that follow.
+    """
+
+    temperature: ForcingUnit = field(
+        validator=validators.instance_of(ForcingUnit),
+        metadata={"description": "Path to the temperature field."},
+    )
+    primary_production: ForcingUnit = field(
+        validator=validators.instance_of(ForcingUnit),
+        metadata={"description": "Path to the primary production field."},
+    )
+    mask: ForcingUnit | None = field(
+        default=None,
+        validator=validators.optional(validators.instance_of(ForcingUnit)),
+        metadata={"description": "Path to the mask field."},
+    )
+    day_length: ForcingUnit | None = field(
+        default=None,
+        validator=validators.optional(validators.instance_of(ForcingUnit)),
+        metadata={"description": "Path to the day length field."},
+    )
+    cell_area: ForcingUnit | None = field(
+        default=None,
+        validator=validators.optional(validators.instance_of(ForcingUnit)),
+        metadata={"description": "Path to the cell area field."},
+    )
+
+    # TODO(Jules): Check that None or both init_cond fields are present in the dataclass
+    initial_condition_production: ForcingUnit | None = field(
+        default=None,
+        validator=validators.optional(validators.instance_of(ForcingUnit)),
+        metadata={"description": "Path to the initial condition production field.", "dims": "Fgroup, <Y, X,> Cohort"},
+    )
+
+    initial_condition_biomass: ForcingUnit | None = field(
+        default=None,
+        validator=validators.optional(validators.instance_of(ForcingUnit)),
+        metadata={"description": "Path to the initial condition biomass field.", "dims": "Fgroup, <Y, X>"},
+    )
+
+    timestep: int | None = field(
+        init=False,
+        default=None,
+        validator=validators.optional(validators.instance_of(int)),
+        metadata={"description": "Common timestep of the fields in day(s)."},
+    )
+
+    resolution: float | tuple[float, float] | None = field(
+        init=False,
+        default=None,
+        metadata={"description": "Common space resolution of the fields as (lat, lon) or both if equals."},
+    )
+
+    def _set_timestep(self: ForcingParameters, forcings: list[ForcingUnit]) -> None:
+        timesteps = {field.timestep for field in forcings if field.timestep is not None}
+        if len(timesteps) != 1:
+            as_dict = dict(zip([field.forcing.name for field in forcings], [field.timestep for field in forcings]))
+            if len(as_dict) != len(timesteps):  # If there are duplicates in the forcing names or None values
+                timesteps = as_dict
+            raise DifferentForcingTimestepError(timesteps)
+        object.__setattr__(self, "timestep", timesteps.pop())
+
+    def _set_resolution(self: ForcingParameters, forcings: list[ForcingUnit]) -> tuple[float, float]:
+        resolutions = {(field.resolution[0], field.resolution[1]) for field in forcings if field.resolution is not None}
+        if len(resolutions) != 1:
+            min_lat = min(lat for lat, _ in resolutions)
+            min_lon = min(lon for _, lon in resolutions)
+            msg = (
+                f"The forcings have different resolutions : {resolutions}."
+                f"\nBe aware that stranges behaviors may occur because minimum resolution is taken : "
+                f"{(min_lat, min_lon)}"
+                f"\nYou can extrapolate the fields to the same resolution using the xarray package."
+            )
+            logger.warning(msg)
+        else:
+            min_lat, min_lon = resolutions.pop()
+        object.__setattr__(self, "resolution", (min_lat, min_lon))
+
+    def _check_units(self: ForcingParameters) -> ForcingUnit:
+        self.temperature.with_units(StandardUnitsLabels.temperature.units, in_place=True)
+        self.primary_production.with_units(StandardUnitsLabels.production.units, in_place=True)
+        if self.day_length is not None:
+            self.day_length.with_units(StandardUnitsLabels.time.units, in_place=True)
+        if self.cell_area is not None:
+            self.cell_area.with_units(StandardUnitsLabels.height.units**2, in_place=True)
+        if self.initial_condition_production is not None:
+            self.initial_condition_production.with_units(StandardUnitsLabels.production.units, in_place=True)
+        if self.initial_condition_biomass is not None:
+            self.initial_condition_biomass.with_units(StandardUnitsLabels.biomass.units, in_place=True)
+
+    def __attrs_post_init__(self: ForcingParameters) -> None:
+        """
+        This method is called after the initialization of the class. It is used to check the consistency of the
+        forcing fields.
+        """
+        forcings = [
+            self.temperature,
+            self.primary_production,
+            self.mask,
+            self.day_length,
+            self.cell_area,
+            self.initial_condition_production,
+            self.initial_condition_biomass,
+        ]
+        forcings = [field for field in forcings if field is not None]
+        self._set_timestep(forcings)
+        self._set_resolution(forcings)
+        self._check_units()
